Fix int-to-double type error in add sale sheet
Dart doesn't implicitly convert int to double in named parameters.
Add explicit .toDouble() on cart item quantity.

+++ lib/presentation/screens/transactions/add_sale_sheet.dart
@@ -455,7 +455,7 @@ class _AddSaleSheetState extends State<AddSaleSheet> {
       items: _cart
           .map((item) => QuickSaleItem(
                 productId: item.product.productId,
-                quantity: item.quantity,
+                quantity: item.quantity.toDouble(),
                 unitPrice: item.product.sellingPrice,
               ))
           .toList(),
